+++ backend/src/main/java/com/calendar/milestone/MilestoneApplication.java
@@ -2,8 +2,7 @@ package com.calendar.milestone;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.context.annotation.Bean;
+
 
 @SpringBootApplication
 public class MilestoneApplication {
